video: tegra: host: use long as timeout variable
wait_event_timeout uses long as the returned timeout value.
However, host drivers have been using "int" variable might
get the sign bit set if the MSB of the returned timeout is
set on LP64 platform (eg. arm64). This causes problem as
negative timeout value indicate errors.

This change fixes the problem by replacing int with long.

Reviewed-on: http://git-master/r/211302
GVS: Gerrit_Virtual_Submit

+++ drivers/video/tegra/host/gk20a/fifo_gk20a.c
@@ -711,7 +711,7 @@ int gk20a_fifo_update_runlist(struct gk20a *g,
 	u32 old_buf, new_buf;
 	u32 chid;
 	u32 count = 0;
-	int remain;
+	long remain;
 	bool pending;
 	u32 ret = 0;
 	u32 token = PMU_INVALID_MUTEX_OWNER_ID;
